fix: 위도경도 자료형 double 변경 및 변수명 lat,lng로 수정

+++ src/main/java/com/ssafy/homer/apartInfo/dto/MapLocaionDto.java
@@ -8,12 +8,12 @@ import lombok.Setter;
 @Getter
 @AllArgsConstructor
 public class MapLocaionDto {
-	private Float startLat;
+	private Double startLat;
 	
-	private Float endLat;
+	private Double endLat;
 	
-	private Float startLng;
+	private Double startLng;
 	
-	private Float endLng;
+	private Double endLng;
 	
 }
